created new class for some methods
